+++ src/models/workouts.rs
@@ -6,13 +6,13 @@ use crate::models::users::User;
 #[diesel(table_name = crate::schema::workouts)]
 #[diesel(check_for_backend(diesel::mysql::Mysql))]
 pub struct Workout {
-    pub id: u64,
-    pub user_id: u64,
+    pub id: i32,
+    pub user_id: i32,
 }
 
 #[derive(Insertable)]
 #[diesel(table_name = crate::schema::workouts)]
 pub struct NewWorkout {
-    pub user_id: u64,
+    pub user_id: i32,
 }
